+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 2.1.0-dev
 - added DummyEventDispatcher
 - fixed method DebugEventDispatcher::dispatched sometimes returning incorrect result
+- EventDispatcher can now log dispatched events
 
 Version 2.0.0
 - removed classes ListenerProvider and PriorityListenerProvider and method ChainListenerProvider::registerProvider